more unit tests and test project refactoring
