+++ MonthlyPremium/Pages/Add.razor
@@ -72,7 +72,7 @@
             </td>
             <td>
                 <InputSelect id="occupation" @bind-Value="_user.Occupation">
-                    <option value="">Select Occupation ...</option>
+                    <option value="">Select Occupation...</option>
                     <option value="WhiteCollar">Author</option>
                     <option value="LightManual">Cleaner</option>
                     <option value="Professional">Doctor</option>
