+++ SolutionTests/SolutionTests.csproj
